[refactor] 重構老師後台的個人資訊頁面
