+++ skilfor/src/pages/TeacherManagePage/components/SelfPage.js
@@ -1,3 +1,7 @@
+import React, { useState, useEffect } from "react";
+import { TEACHER_INFOS } from "../Constant";
+import { sleep } from "../../../utils";
+
 import {
   EditContainer,
   SectionText,
@@ -14,14 +18,33 @@ import {
   EditInput,
 } from "./CourseInfosForm";
 
-function SelfPage({
-  teacherInfos,
-  handleSelfEditClick,
-  handleSelfSubmitClick,
-  isEditingSelf,
-  editSelfContent,
-  setEditSelfContent,
-}) {
+function SelfPage() {
+  const [teacherInfos, setTeacherInfos] = useState(null);
+  //個人資訊是否為編輯狀態
+  const [isEditingSelf, setIsEditingSelf] = useState(false);
+  //編輯個人資料內容
+  const [editSelfContent, setEditSelfContent] = useState(null);
+  //拿取 teacher infos
+  useEffect(() => {
+    async function fetchData() {
+      await sleep(500);
+      setTeacherInfos(TEACHER_INFOS);
+    }
+    fetchData();
+  }, []);
+  //設定預設課程個人編輯 value
+  useEffect(() => {
+    setEditSelfContent(teacherInfos);
+  }, [teacherInfos]);
+  //編輯個人資訊按鈕被按時
+  const handleSelfEditClick = () => setIsEditingSelf(!isEditingSelf);
+  //完成編輯個人資訊按鈕被按時
+  const handleSelfSubmitClick = () => {
+    setIsEditingSelf(false);
+    //將更改後的課程資訊 post 給後端
+    setTeacherInfos(editSelfContent);
+  };
+
   const handleSelfInputChange = (e) => {
     const { id: inputName, value } = e.target;
     switch (inputName) {
@@ -66,56 +89,64 @@ function SelfPage({
         <ItemTop>
           <ItemName>Name</ItemName>
         </ItemTop>
-        <ItemBottom>
-          <ItemValue show={!isEditingSelf}>{teacherInfos.name}</ItemValue>
-          {isEditingSelf && teacherInfos && (
-            <EditInput
-              defaultValue={teacherInfos.name}
-              onChange={handleSelfInputChange}
-              id="name"
-            />
-          )}
-        </ItemBottom>
+        {teacherInfos && (
+          <ItemBottom>
+            <ItemValue show={!isEditingSelf}>{teacherInfos.name}</ItemValue>
+            {isEditingSelf && (
+              <EditInput
+                defaultValue={teacherInfos.name}
+                onChange={handleSelfInputChange}
+                id="name"
+              />
+            )}
+          </ItemBottom>
+        )}
       </FormItemContainer>
       <FormItemContainer show={!isEditingSelf}>
         <ItemTop>
           <ItemName>Avatar</ItemName>
         </ItemTop>
-        <ItemBottom>
-          <ItemValue show={!isEditingSelf}>{teacherInfos.avatar}</ItemValue>
-          {isEditingSelf && teacherInfos && (
-            <EditInput
-              defaultValue={teacherInfos.avatar}
-              onChange={handleSelfInputChange}
-              id="avatar"
-            />
-          )}
-        </ItemBottom>
+        {teacherInfos && (
+          <ItemBottom>
+            <ItemValue show={!isEditingSelf}>{teacherInfos.avatar}</ItemValue>
+            {isEditingSelf && teacherInfos && (
+              <EditInput
+                defaultValue={teacherInfos.avatar}
+                onChange={handleSelfInputChange}
+                id="avatar"
+              />
+            )}
+          </ItemBottom>
+        )}
       </FormItemContainer>
       <FormItemContainer show={!isEditingSelf}>
         <ItemTop>
           <ItemName>Contact Email</ItemName>
         </ItemTop>
-        <ItemBottom>
-          <ItemValue show={!isEditingSelf}>
-            {teacherInfos.contactEmail}
-          </ItemValue>
-          {isEditingSelf && teacherInfos && (
-            <EditInput
-              defaultValue={teacherInfos.contactEmail}
-              onChange={handleSelfInputChange}
-              id="contactEmail"
-            />
-          )}
-        </ItemBottom>
+        {teacherInfos && (
+          <ItemBottom>
+            <ItemValue show={!isEditingSelf}>
+              {teacherInfos.contactEmail}
+            </ItemValue>
+            {isEditingSelf && teacherInfos && (
+              <EditInput
+                defaultValue={teacherInfos.contactEmail}
+                onChange={handleSelfInputChange}
+                id="contactEmail"
+              />
+            )}
+          </ItemBottom>
+        )}
       </FormItemContainer>
       <FormItemContainer show={true}>
         <ItemTop>
           <ItemName>Login Email</ItemName>
         </ItemTop>
-        <ItemBottom>
-          <ItemValue show={true}>{teacherInfos.email}</ItemValue>
-        </ItemBottom>
+        {teacherInfos && (
+          <ItemBottom>
+            <ItemValue show={true}>{teacherInfos.email}</ItemValue>
+          </ItemBottom>
+        )}
       </FormItemContainer>
     </>
   );
